backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
databases[postgresql]==0.9.0
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-jose==3.3.0
alembic==1.14.0
pydantic-settings==2.7.0
pydantic[email]==2.10.4
python-multipart==0.0.20
itsdangerous==2.2.0